resolve unit test hangs
